+++ src/screens/WalletConnect/CallRequest/SignatureRequestContent.js
